requirements.txt
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.2
python-dotenv==1.0.0
python-http-client==3.3.7
sendgrid==6.10.0
starkbank-ecdsa==2.2.0
Werkzeug==2.2.3
